requirements.txt
allure_python_commons==2.13.2
playwright==1.38.0
pytest==7.4.2
